Move L1 labels into lone module
